LayoutTests/fast/selectors/nth-child-of-chained-2-expected.txt

Test chaining :nth-child(An+B) selectors does not aggregate separate selectors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing "testcaseA:nth-child(2n of testcaseA)"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA)').length is 7
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA)')[1].id is "testcase7"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA)')[2].id is "testcase11"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA)')[3].id is "testcase15"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA)')[4].id is "testcase19"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA)')[5].id is "testcase23"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA)')[6].id is "testcase27"
PASS styledElements is ["testcase3","testcase7","testcase11","testcase15","testcase19","testcase23","testcase27"]

Testing "testcaseA:nth-child(2n of [attribute1])"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of [attribute1])').length is 6
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of [attribute1])')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of [attribute1])')[1].id is "testcase11"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of [attribute1])')[2].id is "testcase13"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of [attribute1])')[3].id is "testcase21"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of [attribute1])')[4].id is "testcase23"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of [attribute1])')[5].id is "testcase25"
PASS styledElements is ["testcase3","testcase11","testcase13","testcase21","testcase23","testcase25"]

Testing "testcaseA:nth-child(2n of testcaseA):nth-child(2n of [attribute1])"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA):nth-child(2n of [attribute1])').length is 3
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA):nth-child(2n of [attribute1])')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA):nth-child(2n of [attribute1])')[1].id is "testcase11"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA):nth-child(2n of [attribute1])')[2].id is "testcase23"
PASS styledElements is ["testcase3","testcase11","testcase23"]

Testing "testcaseA:nth-child(2n of testcaseA, [attribute1])"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA, [attribute1])').length is 11
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA, [attribute1])')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA, [attribute1])')[1].id is "testcase5"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA, [attribute1])')[2].id is "testcase7"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA, [attribute1])')[3].id is "testcase9"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA, [attribute1])')[4].id is "testcase11"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA, [attribute1])')[5].id is "testcase13"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA, [attribute1])')[6].id is "testcase21"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA, [attribute1])')[7].id is "testcase23"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA, [attribute1])')[8].id is "testcase25"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA, [attribute1])')[9].id is "testcase27"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(2n of testcaseA, [attribute1])')[10].id is "testcase29"
PASS styledElements is ["testcase3","testcase5","testcase7","testcase9","testcase11","testcase13","testcase21","testcase23","testcase25","testcase27","testcase29"]

PASS successfullyParsed is true

TEST COMPLETE
